portfolio with restaurant.md
This commit marks the initial upload of the portfolio webpage project to GitHub. The project includes the creation of an online restaurant experience within the HOME section, implemented using HTML and CSS. This endeavor represents an exploration into front-end development, combining creativity and technical skill to construct an engaging digital experience. Further iterations and enhancements are anticipated as the project evolves